+++ src/Modules/Workflows/Domain/Steps/Actions/Runners/UnstickPostRunner.php
@@ -49,8 +49,20 @@ class UnstickPostRunner implements StepRunnerInterface
         $this->stepProcessor->executeSafelyWithErrorHandling(
             $step,
             function ($step, $postId) {
-                $postModel = call_user_func($this->expirablePostModelFactory, $postId);
-                $postModel->unstick();
+                /*
+                 * Handle quick-edit saving, otherwise it will override the sticky status at the end of the save
+                 *
+                 * @see https://github.com/publishpress/PublishPress-Future/issues/1204
+                 */
+                if (
+                    defined('DOING_AJAX') && DOING_AJAX &&
+                    isset($_POST['action']) && $_POST['action'] === 'inline-save'
+                ) {
+                    $_POST['sticky'] = false;
+                } else {
+                    $postModel = call_user_func($this->expirablePostModelFactory, $postId);
+                    $postModel->unstick();
+                }
 
                 $nodeSlug = $this->stepProcessor->getSlugFromStep($step);